requirements.txt
Django==1.8.0
#mysqlclient>=1.3.3
django-jsonfield==1.0.1
kombu==3.0.37
futurist
